temporary(autoLD): update for AutoLD's compatibility
The longitudinal pipeline is a temporary improvement for AutoLD. In the near future,
a new version fully generically flexible design must see the light of the day.

A few new estimators been made either as a copy of other (e.g, CFS is a clone copy of CFSPerGroup, MERWAVTIMEMINUS and PLUS been also created but with not proper internal actions as no needed)

This commit is purely temporary. Will need to be soft deleted to be reworked after mini-thesis